MLModel/requirements.txt
Pillow==8.3.2
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
jsonify==0.5
MarkupSafe==2.0.1
nltk==3.6.3
pytesseract==0.3.8
regex==2021.8.28
requests==2.26.0
six==1.16.0
tqdm==4.62.3
urllib3==1.26.6
Werkzeug==2.0.1
gunicorn==20.1.0
